requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.28.2
python-dotenv==0.19.2
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
